Use 16bit ColorSpaces::rgb_to_hsv in "VideoScope" plugin

+++ plugins/videoscope/videoscope.h
@@ -36,7 +36,7 @@
 // Vectorscope HSV axes and labels
 const struct Vectorscope_HSV_axes
 {
-	float  hue;   // angle, degrees
+	int    hue;   // angle, degrees
 	char   label[4];
 	int    color; // label color
 }
